fix: 组织调整，builders和cmd更独立专注于命令行控制流
