+++ src/cmd/builders/routers.ts
@@ -1,8 +1,32 @@
 import fs from "fs";
 import path from "path";
 import rimraf from "rimraf";
-import { GlobalImplements } from "../utils";
-import { InnerRouterOptions, defaultRouterOptions as df } from "../options";
+import { GlobalImplements } from "../../utils";
+
+export interface InnerRouterOptions extends RouterOptions {
+  ctorFolder: string;
+  routerFolder: string;
+}
+
+export interface RouterOptions {
+  /** 是否自动生成2.0的routers，默认：`false` */
+  enabled: boolean;
+  /** 是否强制刷新2.0的routers，默认：`false` */
+  always: boolean;
+  /** 整个项目的url前缀，默认：`'/'` */
+  appRoot: string;
+  /** 生成router文件的文件类型，默认：`'js'` */
+  fileType: "js" | "ts";
+}
+
+export const defaultRouterOptions: InnerRouterOptions = {
+  enabled: false,
+  always: false,
+  ctorFolder: "app/controllers",
+  routerFolder: "app/routers",
+  appRoot: "/",
+  fileType: "js"
+};
 
 interface IRouter {
   [prop: string]: string | IRouter;
@@ -10,12 +34,12 @@ interface IRouter {
 
 export function initRouters(
   {
-    ctorFolder: base = df.ctorFolder,
-    routerFolder: routerBase = df.routerFolder,
-    enabled: open = df.enabled,
-    always = df.always,
-    appRoot: root = df.appRoot,
-    fileType = df.fileType
+    ctorFolder: base = defaultRouterOptions.ctorFolder,
+    routerFolder: routerBase = defaultRouterOptions.routerFolder,
+    enabled: open = defaultRouterOptions.enabled,
+    always = defaultRouterOptions.always,
+    appRoot: root = defaultRouterOptions.appRoot,
+    fileType = defaultRouterOptions.fileType
   }: Partial<InnerRouterOptions>,
   onEnd?: (data: { routers?: IRouter; error?: Error }) => void
 ) {
