+++ src/options/index.ts
@@ -1,6 +1,4 @@
 export * from "./json.options";
 export * from "./render.options";
-export * from "./router.options";
 export * from "./typed-serialize.options";
 export * from "./errors.options";
-export * from "./config-compiler.options";
