+++ src/options/typed-serialize.options.ts
@@ -1,5 +1,5 @@
-import { createOptions } from "../services/Configs";
 import { IStaticTypedResolver } from "../typings/IStaticTypeResolver";
+import { createOptions } from "../typings/IConfigs";
 
 export const STATIC_RESOLVER = createOptions<IStaticTypedResolver>(
   "STATIC_RESOLVER"
