+++ src/services/ConfigReader.ts
@@ -1,21 +1,5 @@
 import { AstroboyContext } from "./AstroboyContext";
 import { Injectable } from "../decorators/injectable";
-import { BaseCompiler } from "../typings/IConfigCompiler";
-
-/**
- * 严格configs预编译接口
- * * 需要实现完整的configs接口
- * * 适合config.default.ts使用
- *
- * @author Big Mogician
- * @export
- * @interface IStrictConfigsCompiler
- * @extends {BaseCompiler<T>}
- * @template T
- */
-export interface IStrictConfigsCompiler<T> extends BaseCompiler<T> {
-  configs(process: NodeJS.Process): T;
-}
 
 /**
  * astroboy基础配置结构
@@ -77,21 +61,6 @@ export interface IAstroboyBaseMiddlewares {
     textLimit?: string;
     strict?: boolean;
   }>;
-}
-
-/**
- * 松散configs预编译接口
- * * 需要实现部分的configs接口
- * * 适合非config.default.ts的文件使用
- *
- * @author Big Mogician
- * @export
- * @interface IConfigsCompiler
- * @extends {BaseCompiler<T>}
- * @template T
- */
-export interface IConfigsCompiler<T> extends BaseCompiler<T> {
-  configs(process: NodeJS.Process): Partial<T>;
 }
 
 /**
