+++ src/typings/IConfigCompiler.ts
@@ -17,6 +17,36 @@ export interface BaseCompiler<T> {
   procedures?(process: NodeJS.Process): string[];
 }
 
+/**
+ * 严格configs预编译接口
+ * * 需要实现完整的configs接口
+ * * 适合config.default.ts使用
+ *
+ * @author Big Mogician
+ * @export
+ * @interface IStrictConfigsCompiler
+ * @extends {BaseCompiler<T>}
+ * @template T
+ */
+export interface IStrictConfigsCompiler<T> extends BaseCompiler<T> {
+  configs(process: NodeJS.Process): T;
+}
+
+/**
+ * 松散configs预编译接口
+ * * 需要实现部分的configs接口
+ * * 适合非config.default.ts的文件使用
+ *
+ * @author Big Mogician
+ * @export
+ * @interface IConfigsCompiler
+ * @extends {BaseCompiler<T>}
+ * @template T
+ */
+export interface IConfigsCompiler<T> extends BaseCompiler<T> {
+  configs(process: NodeJS.Process): Partial<T>;
+}
+
 export interface ConfigDefines {
   modules?: {
     [name: string]: string;
