+++ src/typings/IViewEngine.ts
@@ -1,4 +1,5 @@
-import { Async } from "./IResult";
+/** 表示一个异步过程 */
+export type Async<T> = T | Promise<T>;
 
 export interface IViewEngine {
   render(path: string, configs: any): Async<string>;
